Ajustada a tipagem do atributo quantidade da Classe Ingredients, visto que há possibilidade de atribuir quantidades fracionadas. Também inserida atualização do Diagrama de Classes, v3.

+++ src/models/Ingredient.java
@@ -34,7 +34,7 @@ public class Ingredient {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(float quantity) {
         this.quantity = quantity;
     }
